refactor: CollectionViewController 삭제
- ViewController에서 collectionView 생성
