+++ MapleStoryGuide/MapleStoryGuide/Source/Controller/CoreDataManager.swift
@@ -1,0 +1,8 @@
+//
+//  CoreDataManager.swift
+//  MapleStoryGuide
+//
+//  Created by brad on 2023/03/14.
+//
+
+import Foundation
